requirements.txt
onnx
onnxruntime
numpy
onnx-simplifier
soundfile
librosa
tqdm